+++ packages/gi-assets-advance/src/components/Redo/index.tsx
@@ -1,17 +1,7 @@
 import Component from './Component';
+import info from './info';
 import registerMeta from './registerMeta';
-
 /**   index.md 中解析得到默认值，也可用户手动修改 */
-
-const info = {
-  id: 'Redo',
-  name: '重做',
-  desc: '重做',
-  cover: 'http://xxxx.jpg',
-
-  category: 'canvas-interaction',
-  type: 'GIAC',
-};
 
 export default {
   info,
