+++ packages/gi-assets-advance/src/components/Redo/info.ts
@@ -1,0 +1,10 @@
+const info = {
+  id: 'Redo',
+  name: '重做',
+  desc: '重做，画布进入下一步状态',
+  cover: 'http://xxxx.jpg',
+  category: 'canvas-interaction',
+  type: 'GIAC',
+  icon: 'icon-redo',
+};
+export default info;
