+++ packages/gi-assets-advance/src/components/Undo/info.ts
@@ -1,0 +1,10 @@
+const info = {
+  id: 'Undo',
+  name: '撤销',
+  desc: '撤销，画布返回上一步状态',
+  cover: '撤销',
+  category: 'canvas-interaction',
+  type: 'GIAC',
+  icon: 'icon-undo',
+};
+export default info;
